Crate origen_metal

Re-exports

pub extern crate cfg_if;

pub extern crate config;

pub extern crate dialoguer;

pub extern crate futures;

pub extern crate glob;

pub extern crate indexmap;

pub extern crate lazy_static;

pub extern crate minijinja;

pub extern crate octocrab;

pub extern crate reqwest;

pub extern crate serde_json;

pub extern crate tera;

pub extern crate toml_edit;

pub use framework::sessions::sessions;

pub use framework::sessions::Sessions;

pub use framework::typed_value::Map as TypedValueMap;

pub use framework::typed_value::TypedValue;

pub use framework::typed_value::TypedValueVec;

pub use frontend::with_frontend;

pub use frontend::with_optional_frontend;

pub use utils::file;

pub use utils::outcome::Outcome;

pub use utils::outcome::OutcomeState;

pub use utils::outcome::OutcomeSubTypes;

pub use utils::outcome::OutcomeSubtypes;

pub use framework::users::users::add_user;

pub use framework::users::users::clear_current_user;

pub use framework::users::users::get_current_user_email;

pub use framework::users::users::get_current_user_home_dir;

pub use framework::users::users::get_current_user_id;

pub use framework::users::users::get_initial_user_id;

pub use framework::users::users::require_current_user_email;

pub use framework::users::users::require_current_user_home_dir;

pub use framework::users::users::require_current_user_id;

pub use framework::users::users::set_current_user;

pub use framework::users::users::try_lookup_and_set_current_user;

pub use framework::users::users::try_lookup_current_user;

pub use framework::users::users::users;

pub use framework::users::users::users_mut;

pub use framework::users::users::with_current_user;

pub use framework::users::users::with_current_user_session;

pub use framework::users::users::with_user;

pub use framework::users::users::with_user_mut;

pub use framework::users::users::with_user_or_current;

pub use framework::users::users::with_users;

pub use framework::users::users::with_users_mut;

pub use framework::users::user::add_dataset_to_user;

pub use framework::users::user::register_dataset_with_user;

pub use framework::users::user::with_user_dataset;

pub use framework::users::user::with_user_dataset_mut;

pub use framework::users::user::with_user_hierarchy;

pub use framework::users::user::with_user_motive_or_default;

pub use framework::users::users::unload as unload_users;

pub use utils::os::on_linux as running_on_linux;

pub use utils::os::on_windows as running_on_windows;

Modules

_utility

ast

built_info

framework

The framework module contains functionality that is closely related to implementing a a particular framework feature. While still generic in nature, they are not so low-level as the functions found in the utils module which could be used in a variety of different contexts. The functionality in this module assumes that it will be used to implement something very similar to Origen’s implementation of the given feature.

frontend

macros

prelude

prog_gen

stil

utils

Macros

backend_expect

backend_fail

bail

current_func

Get the caller name. Taken from this SO answer: https://stackoverflow.com/a/63904992/8533619

display

display_cyan

display_cyanln

display_green

display_greenln

display_red

display_redln

display_yellow

display_yellowln

displayln

error

exit_error

Exit the origen process with a failing exit code and a big ERROR banner

exit_fail

Exit the origen process with a failing exit code and a big FAIL banner

exit_pass

Exit the origen process with a passing exit code and a big PASS banner

exit_success

Exit the origen process with a passing exit code and a big SUCCESS banner

hashmap

into_aes_gcm_generic_array

log_debug

log_deprecated

log_error

log_info

log_success

log_trace

log_warning

new_cmd

node

scrub_path

Create a new pathbuf, rebuilt from components to resolve OS path differences when displaying paths possibly built from “/”, “", or “\” used interchangeably

trace

Structs

Error

FLOW

This is analogous to the TEST for test program duration, it provides a similar API for pushing nodes to the current flow, FLOW.push(my_node), etc.

FRONTEND

LOGGER

PROG_GEN_CONFIG

SESSIONS

USERS

VERSION

Traits

Context

Functions

unload

Type Aliases

RefStrAble

Result