origen_metal

Module _utility

Modules

file_utils

Functions

bytes_from_str_of_bytes

For convenience in converting to/from configs, allow bytes to be converted to a string representation of byte values (NOT characters themselves). This allows for invalid UTF-8 characters to still be encoded in a string.

resolve_os_str

str_from_byte_array

str_to_bool

unsorted_dedup

validate_input_list

Checks the given values of a vector against an enumerated set of accepted values. Optionally, check for duplicate items as well.