origen_metal::ast

Trait Processor

pub trait Processor<T> {
    // Provided methods
    fn on_node(&mut self, _node: &Node<T>) -> Result<Return<T>> { ... }
    fn on_end_of_block(&mut self, _node: &Node<T>) -> Result<Return<T>> { ... }
    fn on_processed_node(&mut self, _node: &Node<T>) -> Result<Return<T>> { ... }
}

Provided Methods

fn on_node(&mut self, _node: &Node<T>) -> Result<Return<T>>

fn on_end_of_block(&mut self, _node: &Node<T>) -> Result<Return<T>>

This will be called at the end of processing every node which has children. The node which is about to be closed is provided in the arguments. Note that you should probably never return a derivative of the given node here, it should either be None or a new node(s)

fn on_processed_node(&mut self, _node: &Node<T>) -> Result<Return<T>>