origen_metal::framework

Module reference_files

This manages how generated files are saved as references. When new or changed files are generated a save ref is saved on disk which will allow a future execution of the ‘origen save_ref’ command to determine the source and destination of the file to be saved as a the new reference version. Save refs are stored as single files per reference instead of a single file containing all references or the log as was used in O1. A single file is not used to avoid contentions when multiple generation jobs are running in parallel on the LSF.

Functions

apply_all_changed_refs

Apply all previously created changed file references (copys the new versions of the files to their reference file counterparts)

apply_all_new_refs

Apply all previously created new file references (copys the new versions of the files to their reference file counterparts)

apply_ref

Apply a previously created changed/new file reference (copys the new version of the file to its reference file counterpart)

clear_save_refs

Clear (delete) all existing changed and new file references that have not been applied yet.

create_changed_ref

Create a new record of a changed file, providing a key with which to reference it by in future, a path to the new file and a path to the reference file.

create_new_ref

Create a new record of a new file, providing a key with which to reference it by in future, a path to the new file and a path to the reference file (which doesn’t exist yet).

set_save_ref_dir

This must be called before all other APIs to set the directory where temporary files can be stored to keep track of which files have changed, are new etc. The given location should not be added to revision control.