origen_metal::framework::users

Struct Data

pub struct Data {

    pub dataset_name: String,
    pub password: Option<String>,
    pub name: Option<String>,
    pub username: Option<String>,
    pub first_name: Option<String>,
    pub last_name: Option<String>,
    pub display_name: Option<String>,
    pub other: TypedValueMap,
    pub home_dir: Option<PathBuf>,
    pub name_tried: bool,
    pub email: Option<String>,
    pub email_tried: bool,
    pub authenticated: bool,
    pub populated: bool,
    pub populate_attempted: bool,
    pub roles: Vec<Roles>,
    /* private fields */

}

Struct to hold user data for a given dataset. At least in the backend, each data instance is independent with an dataset-dependent features (e.g., username, password, retrieval) handled by the owning User struct.

Fields
dataset_name: Stringpassword: Option<String>name: Option<String>username: Option<String>first_name: Option<String>last_name: Option<String>display_name: Option<String>other: TypedValueMaphome_dir: Option<PathBuf>name_tried: boolemail: Option<String>email_tried: boolauthenticated: boolpopulated: boolpopulate_attempted: boolroles: Vec<Roles>
Implementations

impl Data

pub fn new(dataset_name: &str, config: &DatasetConfig) -> Self

pub fn get_display_name(&self) -> Option<String>

pub fn require_home_dir(&self, parent: &User) -> Result<PathBuf>

pub fn password_key(&self) -> String

pub fn clear_cached_password(&mut self, parent: &User) -> Result<()>

pub fn set_should_validate_password(&mut self, new_val: Option<bool>)

pub fn config(&self) -> &DatasetConfig

pub fn password_needs_validation(&self) -> bool

pub fn require_data_source_for( &self, op: &str, user_id: &str, ) -> Result<(&str, &str)>

pub fn populate_succeeded(&self) -> Option<bool>

pub fn populate_failed(&self) -> Option<bool>

pub fn has_empty_populate_config(&self) -> bool

pub fn should_auto_populate(&self) -> bool

pub fn populate( user: &User, ds_name: &str, repopulate: bool, continue_on_error: bool, stop_on_failure: bool, ) -> Result<Option<Outcome>>