origen_metal::framework::users::users

Struct Users

pub struct Users { /* private fields */ }

Implementations

impl Users

pub fn get_current_id(&self) -> Result<Option<&String>>

pub fn get_current_user(&self) -> Result<Option<&User>>

pub fn get_current_user_mut(&mut self) -> Result<Option<&mut User>>

pub fn get_user(&self, id: &str) -> Result<Option<&User>>

pub fn get_user_mut(&mut self, id: &str) -> Result<Option<&mut User>>

pub fn get_initial_id(&self) -> Result<Option<&String>>

pub fn get_initial_user(&self) -> Result<Option<&User>>

pub fn current_id(&self) -> Result<&str>

pub fn current_user(&self) -> Result<&User>

pub fn current_user_mut(&mut self) -> Result<&mut User>

pub fn user(&self, u: &str) -> Result<&User>

pub fn user_mut(&mut self, u: &str) -> Result<&mut User>

pub fn users(&self) -> &IndexMap<String, User>

pub fn set_current_user(&mut self, id: &str) -> Result<bool>

pub fn clear_current_user(&mut self) -> Result<bool>

pub fn try_lookup_current_user() -> Result<String>

pub fn try_lookup_and_set_current_user() -> Result<(String, Option<Option<PopulateUserReturn>>)>

pub fn add_user( id: &str, auto_populate: Option<bool>, ) -> Result<Option<PopulateUserReturn>>

pub fn add_user_option_less(id: &str) -> Result<Option<PopulateUserReturn>>

pub fn remove(&mut self, id: &str) -> Result<bool>

pub fn initial_id(&self) -> Result<&str>

pub fn initial_user(&self) -> Result<&User>

pub fn unload(&mut self, _continue_on_user_unload_fail: bool) -> Result<()>

pub fn default_datasets(&self) -> &IndexMap<String, DatasetConfig>

pub fn require_default_dataset(&self, dataset: &str) -> Result<&DatasetConfig>

pub fn default_datakeys(&self) -> Vec<&String>

pub fn override_default_dataset( &mut self, new_ds: &str, new_ds_config: DatasetConfig, ) -> Result<()>

Overrides the default datakey/dataset, removing the origen default and replacing it with this one. This can only be done prior to adding any users or auxillary datasets, otherwise an error is thrown

pub fn default_data_lookup_hierarchy(&self) -> &Vec<String>

pub fn set_default_data_lookup_hierarchy( &mut self, hierarchy: Vec<String>, ) -> Result<()>

pub fn register_default_dataset( &mut self, name: &str, config: DatasetConfig, ) -> Result<()>

Registers a new default dataset but does not add it to the hierarchy

pub fn add_default_dataset( &mut self, name: &str, config: DatasetConfig, as_topmost: bool, ) -> Result<()>

Registers a new default dataset and adds it to the hierarchy. ‘as_topmost’ = true -> inserts at the beginning (highest) hierarchy ‘as_topmost’ = false -> inserts at the lowest point in the hierarchy

pub fn motive_mapping(&self) -> &IndexMap<String, String>

pub fn add_motive( &mut self, motive: String, dataset: String, replace_existing: bool, ) -> Result<Option<String>>

pub fn dataset_for(&self, motive: &str) -> Result<Option<&String>>

pub fn populate( &self, repopulate: bool, continue_on_error: bool, stop_on_failure: bool, ) -> Result<PopulateUsersReturn>

pub fn default_auto_populate(&self) -> &Option<bool>

pub fn set_default_auto_populate(&mut self, set_to: Option<bool>)

pub fn default_session_config(&self) -> &SessionConfig

pub fn default_session_config_mut(&mut self) -> &mut SessionConfig

pub fn password_encryption_key(&self) -> &str

pub fn password_encryption_nonce(&self) -> &str

pub fn default_should_validate_passwords(&self) -> &Option<bool>

pub fn set_default_should_validate_passwords(&mut self, set_to: Option<bool>)

pub fn default_password_cache_option(&self) -> &Option<PasswordCacheOptions>

pub fn set_default_password_cache_option<P: TryInto<PasswordCacheOptions, Error = Error>>( &mut self, set_to: P, ) -> Result<()>

pub fn default_roles(&self) -> Result<&Vec<String>>

pub fn set_default_roles<S: AsRef<str>>(&mut self, roles: &Vec<S>) -> Result<()>

pub fn clear_default_roles(&mut self) -> Result<()>

pub fn roles(&self) -> Result<Vec<String>>

Return all roles

pub fn users_by_role( &self, filter: Option<&dyn Fn(&User, &String) -> bool>, ) -> Result<HashMap<String, Vec<String>>>