origen_metal::prog_gen

Enum SupportedTester

pub enum SupportedTester {
    ALL,
    V93K,
    V93KSMT7,
    V93KSMT8,
    IGXL,
    J750,
    ULTRAFLEX,
    SIMULATOR,
    Extension(u32),
}

Variants

ALL

Generally, the absence of an optional SupportedTester value means all testers, but this can also be used to indicate that whenever a SupportedTester value is required

V93K

Indicates support for all versions of SMT on the V93K

V93KSMT7

Indicates support for V93K SMT7 only, i.e. not SMT8

V93KSMT8

Indicates support for V93K SMT8 only, i.e. not SMT7

IGXL

Indicates support for all IGXL-based testers

J750

ULTRAFLEX

SIMULATOR

Extension(u32)

Opaque extension slot for proprietary tester implementations in downstream crates.

The u32 tag is chosen by the downstream crate and is never interpreted by origen_metal itself. Extension variants are intentionally excluded from SupportedTester::all_names() and std::str::FromStr — they are not user-facing tester names.

Example (downstream crate)

use origen_metal::prog_gen::SupportedTester;

// Pick any u32 tag that is meaningful to your crate.
const MY_TESTER: SupportedTester = SupportedTester::Extension(0x4D595452); // "MYTR"

// Use it anywhere SupportedTester is required:
let model = origen_metal::prog_gen::Model::new(MY_TESTER);

Implementations

impl SupportedTester

pub fn all_names() -> Vec<String>

Returns the names of all available testers

pub fn new(name: &str) -> OrigenResult<Self>

pub fn is_compatible_with(&self, tester: &SupportedTester) -> bool

Returns true if the given tester is equal to self or if the given tester is a derivative of self (see is_a_derivative_of()).

pub fn is_a_derivative_of(&self, tester: &SupportedTester) -> bool

Returns true if self is a derivative of the given tester. For example if the given tester is IGXL, then both the J750 and the ULTRAFLEX would be considered derivatives of it. Note that true is only returned for derivatives, if the given tester is the same as self then false will be returned. Use is_compatible_with() if you also want to consider an exact match as true.