origen_metal::prog_gen

Module flow_api

Functions

assign_test_to_invocation

bin

Bin out the DUT

continue_on_fail

define_bin

define_igxl_resource

Define a row in an IG-XL resource worksheet.

define_pattern_group

define_sub_test

define_test

Defines a new test in the AST, returning its ID. A test must be initially defined before attributes can be set on it. It won’t actually appear in the test flow until it is added to if via add_test()

define_test_collection_item

define_test_invocation

Defines a new test invocation in the AST, returning its ID. A test invoration must be initially defined before attributes can be set on it. It won’t actually appear in the test flow until it is added to if via add_test()

end_block

execute_cz_test

Cz the given test (or invocation) from the current flow

execute_test

Execute the given test (or invocation) from the current flow

execute_test_str

Execute the given test (or invocation) from the current flow, where the test is a string that will be rendered verbatim to the flow - no linkage to an actual test object will be checked or inserted by Origen

flow_description

flow_name_override

log

push_pattern_to_group

render

Renders the given string directly to the test flow

set_default_flag_state

set_flag

set_igxl_resources_filename

set_namespace

set_resources_filename

set_test_attr

Set an attribute of either a test or a test invocation

set_test_limit

set_volatile_flags

Mark the given flags as volatile, meaning their state could be changed by tests and therefore should not be optimized out of the flow

set_wait_flags

[IGXL only] Set the given wait flags on the given test instance

start_bypass_sub_flows

Contained flows will be bypassed

start_condition

start_group

Used to model flow groups, IG-XL test instance groups, etc.

start_on_failed

Start an on-failed block (events to run if the given test or group failed), the returned reference should be retained and passed to end_block

start_on_passed

Start an on-passed block (events to run if the given test or group passed), the returned reference should be retained and passed to end_block

start_resources

Start a resources block, contained tests will not appear in the flow but the test definitions will appear in the generated test program (e.g. in the test instances sheet)

start_sub_flow

Start a sub-flow, the returned reference should be retained and passed to end_block

start_uniqueness

Contained test names and similar will the use the given uniqueness option