origen_metal::prog_gen

Function process_flow

pub fn process_flow(
    flow: &AST<PGM>,
    model: Model,
    tester: SupportedTester,
    validate: bool,
) -> Result<(Node<PGM>, Model)>

Processes the given flow AST so that it is ready to generate the flow for the given tester, optionally validating it first

use origen_metal::FLOW;
use origen_metal::prog_gen::Model;
use origen_metal::prog_gen::SupportedTester;
use origen_metal::prog_gen::process_flow;

FLOW.with_all_flows(|flows| {
    let mut model = Model::new(SupportedTester::V93KSMT7);
    for (name, flow) in flows {
        let ast;
        (ast, model) = process_flow(flow, model, SupportedTester::V93KSMT7, true)?;
    }
    Ok(())
});