origen_metal::prog_gen

Struct Model

pub struct Model {
    pub tester: SupportedTester,
    pub tests: IndexMap<usize, Test>,
    pub test_invocations: IndexMap<usize, Test>,
    pub test_instance_groups: IndexMap<String, Vec<usize>>,
    pub test_collection_items: IndexMap<usize, TestCollectionItem>,
    pub sub_tests: Vec<SubTest>,
    pub patterns: Vec<Pattern>,
    pub variables: Vec<Variable>,
    pub flows: IndexMap<String, Flow>,
    pub pattern_collections: IndexMap<String, Vec<usize>>,
    pub variable_collections: IndexMap<String, Vec<usize>>,
    pub custom_data: IndexMap<String, String>,
    /* private fields */
}

The test program model contains tests, test invocations, patterns, bins, etc. that have been extracted from a flow AST into a generic data structure that can be consumed by all tester targets.

Fields
tester: SupportedTestertests: IndexMap<usize, Test>

Test objects, stored by their internal ID. These map to test instances for IG-XL and test methods for V93K.

test_invocations: IndexMap<usize, Test>

Test invocation objects, stored by their internal ID. These map to test flow lines for IG-XL and test suites for V93K.

test_instance_groups: IndexMap<String, Vec<usize>>

IG-XL test-instance group name to member test IDs.

test_collection_items: IndexMap<usize, TestCollectionItem>

Collection item objects referenced by test methods and nested collection items.

sub_tests: Vec<SubTest>

Tests can store a single limit, but if a test has multiple limits then they are represented as sub-tests

patterns: Vec<Pattern>

All pattern references made in the test program, flows and pattern_collections make reference to these via their ID (their vector index number)

variables: Vec<Variable>

All variable references made in the test program, flows and variable_collections make reference to these via their ID (their vector index number)

flows: IndexMap<String, Flow>pattern_collections: IndexMap<String, Vec<usize>>variable_collections: IndexMap<String, Vec<usize>>custom_data: IndexMap<String, String>

Generic key-value metadata for use by importers and exporters. Provides a place to store format-level data (version strings, schema tags, etc.) that does not fit the structured model fields. Processors ignore this field; it is for import/export tooling only.

Implementations

impl Model

pub fn new(tester: SupportedTester) -> Self

impl Model

pub fn set_resources_filename(&mut self, name: String, kind: &ResourcesType)

pub fn patterns_from_ids( &self, ids: &Vec<usize>, sort: bool, uniq: bool, ) -> Vec<&Pattern>

pub fn variables_from_ids( &self, ids: &Vec<usize>, sort: bool, uniq: bool, ) -> Vec<&Variable>

pub fn select_flow(&mut self, name: &str) -> Result<()>

Set the current flow (default flow operated on by some of the model’s methods), returns an error if the model doesn’t contain a flow with the given name

pub fn create_flow(&mut self, name: &str) -> Result<()>

Creates a new flow within the model and selects it as the current flow. An error will be returned if a flow of the given name already exists.

pub fn get_flow(&self, name: Option<&str>) -> Option<&Flow>

Get a reference to the current or given flow

pub fn get_flow_mut(&mut self, name: Option<&str>) -> &mut Flow

Get a mutable reference to the current or given flow, will create it if it doesn’t exist yet

pub fn current_pattern_collection_name(&self) -> &str

pub fn get_pattern_collection(&self, name: Option<&str>) -> Option<&Vec<usize>>

Get a reference to the current or given pattern collection

pub fn get_pattern_collection_mut( &mut self, name: Option<&str>, ) -> &mut Vec<usize>

Get a mutable reference to the current or given resource, will create it if it doesn’t exist yet

pub fn record_pattern_reference( &mut self, path: String, pattern_type: Option<PatternType>, reference_type: Option<PatternReferenceType>, )

Record a pattern reference and allocate to the current flow and pattern collection

pub fn current_variable_collection_name(&self) -> &str

pub fn get_variable_collection(&self, name: Option<&str>) -> Option<&Vec<usize>>

Get a reference to the current or given variable collection

pub fn get_variable_collection_mut( &mut self, name: Option<&str>, ) -> &mut Vec<usize>

Get a mutable reference to the current or given resource, will create it if it doesn’t exist yet

pub fn record_variable_reference( &mut self, name: String, variable_type: VariableType, operation: VariableOperation, )

Record a variable reference and allocate to the current flow and variable collection

pub fn add_test_from_template( &mut self, id: usize, name: String, tester: &SupportedTester, template_name: &str, library_name: Option<&str>, ) -> Result<()>

Create a new test within the model from the given template reference. An error will be returned if the given template can not be found, or if a test alraedy exists with the given ID.

pub fn add_test_to_instance_group(&mut self, group: &str, test_id: usize)

pub fn test_instance_group_name(&self, test_id: usize) -> Option<&str>

pub fn add_test_invocation( &mut self, id: usize, name: String, tester: &SupportedTester, ) -> Result<()>

Create a new test invocation within the model from the given tester reference. An error will be returned if a test invocation alraedy exists with the given ID.

pub fn assign_test_to_inv( &mut self, inv_id: usize, test_id: usize, ) -> Result<()>

Assign the given test to the given invocation, returns an error if neither exists. Currently, no error will be raised if a test is already assigned to the invocation, it will be replaced.

pub fn add_sub_test( &mut self, test_id: usize, name: String, number: Option<usize>, lo_limit: Option<Limit>, hi_limit: Option<Limit>, ) -> Result<()>

pub fn set_test_attr( &mut self, id: usize, name: &str, value: Option<ParamValue>, allow_missing: bool, ) -> Result<()>

Set the value of the given test attribute. If the given ID refers to a test invocation then both the invocation and the test will be checked for a matching attribute. Currently, if no matching attribute is found then nothing happens. An error is returned if the test doesn’t exist or if the value is the wrong type for the given parameter. Calling with value = None will cause all existing settings for the given attribute to be removed.

pub fn add_test_collection_item( &mut self, parent_id: usize, item_id: usize, collection_name: &str, instance_id: &str, allow_missing: bool, ) -> Result<()>