origen_metal

Module stil

Re-exports

pub use nodes::STIL;

Modules

nodes

processors

Structs

Parser

A configurable STIL parser.

Enums

Alignment

Base

Selector

SignalType

State

Termination

Functions

from_file

from_file_ignore_includes

from_file_with_options

Parse the given STIL file, using the given load path to resolve any include statements that are encountered. Include files can optionally be renamed using the rename argument, which is a map of the original include file name to the new name.

from_str