origen_metal::utils::revision_control

Trait RevisionControlAPI

pub trait RevisionControlAPI: Debug {

    // Required methods
    fn populate(&self, version: &str) -> Result<()>;
    fn checkout(
        &self,
        force: bool,
        path: Option<&Path>,
        version: &str,
    ) -> Result<bool>;
    fn revert(&self, path: Option<&Path>) -> Result<()>;
    fn status(&self, path: Option<&Path>) -> Result<Status>;
    fn tag(
        &self,
        tagname: &str,
        force: bool,
        message: Option<&str>,
    ) -> Result<()>;
    fn init(&self) -> Result<Outcome>;
    fn is_initialized(&self) -> Result<bool>;
    fn checkin(
        &self,
        files_or_dirs: Option<Vec<&Path>>,
        msg: &str,
        dry_run: bool,
    ) -> Result<Outcome>;
    fn system(&self) -> &str;

    // Provided methods
    fn remote_url(&self) -> Result<String> { ... }
    fn current_branch(&self) -> Result<String> { ... }
    fn tag_exists(&self, _tagname: &str) -> Result<bool> { ... }
    fn confirm_latest_ref(&self, _branch: &str) -> Result<(bool, [String; 2])> { ... }

}

Defines a common minimum API that all revision control system drivers should support

Required Methods

fn populate(&self, version: &str) -> Result<()>

Initially populate the local directory with the remote, this is equivalent to a ‘git clone’ or a ‘dssc pop’ operation. A progress instance will be returned indicating how many objects were fetched. A callback can be given which will be called periodically if the caller wants to be updated on the progress during the operation.

fn checkout( &self, force: bool, path: Option<&Path>, version: &str, ) -> Result<bool>

Checkout the given version of the repository. If force is false then local modifications will be preserved and merged with any upstream changes. If merge conflicts are encountered then this method will return Ok(true)

fn revert(&self, path: Option<&Path>) -> Result<()>

Reverts any local changes. Supplying a path to a directory may be supported to limit the results to files that fall withing the given directory.

fn status(&self, path: Option<&Path>) -> Result<Status>

Returns a Status object which contains lists of all files which have local modifications. Supplying a path to a directory may be supported to limit the results to files that fall withing the given directory.

fn tag(&self, tagname: &str, force: bool, message: Option<&str>) -> Result<()>

Tag the view in the local workspace. A tag message can be supplied, but this may or may not be applied to the repo depending on whether the underlying system supports it. Supplying force: true will replace any existing tag with the same name.

fn init(&self) -> Result<Outcome>

Initialize a new local workspace at path, pointing to the given location Returns true if a new workspace was created, false if the workspace already existed (no action), or an error.

fn is_initialized(&self) -> Result<bool>

Indicate if the path of the RC driver is initialized

fn checkin( &self, files_or_dirs: Option<Vec<&Path>>, msg: &str, dry_run: bool, ) -> Result<Outcome>

fn system(&self) -> &str

Provided Methods

fn remote_url(&self) -> Result<String>

fn current_branch(&self) -> Result<String>

fn tag_exists(&self, _tagname: &str) -> Result<bool>

fn confirm_latest_ref(&self, _branch: &str) -> Result<(bool, [String; 2])>